+++ clang/lib/Sema/SemaStmt.cpp
@@ -717,12 +717,11 @@ bool Sema::checkMustTailAttr(const Stmt *St, const Attr &MTA) {
     return false;
   }
 
-  if (const FunctionDecl *CalleeDecl = CE->getDirectCallee()) {
-    if (CalleeDecl->hasAttr<NotTailCalledAttr>()) {
-      Diag(St->getBeginLoc(), diag::err_musttail_conflicts_with_not_tail_called)
-          << &MTA;
-      return false;
-    }
+  if (const FunctionDecl *CalleeDecl = CE->getDirectCallee();
+     CalleeDecl && CalleeDecl->hasAttr<NotTailCalledAttr>()) {
+      Diag(St->getBeginLoc(), diag::err_musttail_mismatch)
+      << &MTA;
+     return false;
   }
 
   if (const auto *EWC = dyn_cast<ExprWithCleanups>(E)) {